# docs/swagger.yaml
definitions:
  errcode.Error:
    type: object
...
  model.RequestResetInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  model.ResetPasswordSubmitInput:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
...
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - firstName
    - gender
    - lastName
    - password
    - username
    type: object
...
  model.UserToken:
    properties:
      accessToken:
        type: string
      expired:
        type: string
    type: object
...
info:
  contact: {}
  description: health master open api
  termsOfService: http://swagger.io/terms/
  title: health master api
  version: "1.0"
paths:
...
  /api/v1/password-resets:
    post:
      consumes:
      - application/json
      description: send a password reset link to the user's email
      parameters:
      - description: email address
        in: body
        name: RequestResetInput
        required: true
        schema:
          $ref: '#/definitions/model.RequestResetInput'
      produces:
      - application/json
      responses:
        "200":
          description: password reset link sent
          schema:
            $ref: '#/definitions/errcode.Error'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: create a password reset request
      tags:
      - password-reset
  /api/v1/password-resets/{token}:
    get:
      consumes:
      - application/json
      description: verify the validity of a password reset token
      parameters:
      - description: password reset token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: masked email
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid or expired token
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: verify password reset token
      tags:
      - password-reset
    put:
      consumes:
      - application/json
      description: reset user password with a valid password reset token
      parameters:
      - description: password reset token
        in: path
        name: token
        required: true
        type: string
      - description: new password and confirmation
        in: body
        name: ResetPasswordSubmitInput
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordSubmitInput'
      produces:
      - application/json
      responses:
        "200":
          description: password reset successfully
          schema:
            $ref: '#/definitions/errcode.Error'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: reset password
      tags:
      - password-reset
...
  /api/v1/tmptoken:
    post:
      consumes:
      - application/json
      description: get a temporary token (1 minute expiry)
      parameters:
      - description: access token
        in: query
        name: accessToken
        type: string
      - default: Bearer <Add access token here>
        description: Authorization token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.UserToken'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: get tmp token
      tags:
      - token
  /api/v1/token:
    get:
      consumes:
      - application/json
      description: parse a exsited token
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: parse a token
      tags:
      - token
    post:
      consumes:
      - application/json
      description: create a new token
      parameters:
      - description: username
        in: header
        name: username
        required: true
        type: string
      - description: password
        in: header
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.UserToken'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: create a token
      tags:
      - token
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"